using unsinged without crashing in disassembly.

+++ src/gtirb-decoder/arch/X64Loader.cpp
@@ -140,7 +140,7 @@ std::optional<relations::Operand> X64Loader::build(const cs_x86_op& CsOp)
                                        registerName(CsOp.mem.index),
                                        CsOp.mem.scale,
                                        CsOp.mem.disp,
-                                       CsOp.size * 8};
+                                       static_cast<uint64_t>(CsOp.size) * 8};
             return I;
         }
         case X86_OP_INVALID:
